various updates

+++ start.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 
 
-eval $(conda shell.bash hook)
-conda activate cmulab
-source activate cmulab
+PORT=${1:-8088}
+
+. /opt/conda/etc/profile.d/conda.sh
+[[ $CONDA_DEFAULT_ENV == "base" ]] || eval $(command conda shell.bash hook)
+conda activate cmulab 2>/dev/null
+[[ $CONDA_DEFAULT_ENV == "cmulab" ]] || source activate cmulab
+[[ $CONDA_DEFAULT_ENV == "cmulab" ]] || { echo "Couldn't activate conda env cmulab. Exiting!"; exit 1; }
+echo "conda activate $CONDA_DEFAULT_ENV"
+
 set -x
 
 pkill -f rqworker
@@ -17,11 +23,16 @@ export PYTHONUNBUFFERED=1
 export GOOGLE_APPLICATION_CREDENTIALS
 export OCR_POST_CORRECTION
 export EMAIL_HOST_PASSWORD
-nohup python -u manage.py runserver 0.0.0.0:8088 --verbosity 2 &> log.runserver_8088 &
-nohup redis-server redis.conf &> log.redis-server &
-sleep 1
+
+timestamp=$(date +"%Y%m%d%H%M%S")
+mkdir -p logs/${timestamp}
+mv logs/log.* logs/${timestamp}
+
+nohup python -u manage.py runserver 0.0.0.0:${PORT} --verbosity 2 &> logs/log.runserver_${PORT} &
+nohup redis-server redis.conf &> logs/log.redis-server &
+sleep 5
 
 NUM_JOBS=3
 for n in $(seq 1 $NUM_JOBS); do
-    nohup python -u manage.py rqworker default &> log.rqworker_default_${n} &
+    nohup python -u manage.py rqworker default &> logs/log.rqworker_default_${n} &
 done
